requirements.txt
numpy
matplotlib
pytest
rasterio
requests
pyyaml
geopandas
matplotlib.colors
scipy
shapely